changefeedccl: Cleanup and refactor sink construction.
Cleanup sink construction factory method to take fewer arguments.
Refactor and move non-trivial sinks (kafka and sql) into standalone
files.

Release Notes: none
