+++ pkg/ccl/changefeedccl/sink.go
@@ -10,42 +10,20 @@ package changefeedccl
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
-	gosql "database/sql"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
-	"hash"
-	"hash/fnv"
 	"net/url"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 
-	"github.com/Shopify/sarama"
-	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/security"
-	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/builtins"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/storage/cloud"
 	"github.com/cockroachdb/cockroach/pkg/util/bufalloc"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
-	"github.com/cockroachdb/cockroach/pkg/util/humanizeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
-	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/errors"
-	"github.com/cockroachdb/logtags"
 )
 
 // TopicDescriptor describes topic emitted by the sink.
@@ -62,6 +40,8 @@ type TopicDescriptor interface {
 
 // Sink is an abstraction for anything that a changefeed may emit into.
 type Sink interface {
+	// Dial establishes connection to the sink.
+	Dial() error
 	// EmitRow enqueues a row message for asynchronous delivery on the sink. An
 	// error may be returned if a previously enqueued message has failed.
 	EmitRow(ctx context.Context, topic TopicDescriptor, key, value []byte, updated hlc.Timestamp) error
@@ -80,189 +60,79 @@ type Sink interface {
 
 func getSink(
 	ctx context.Context,
-	sinkURI string,
-	srcID base.SQLInstanceID,
-	opts map[string]string,
-	targets jobspb.ChangefeedTargets,
-	settings *cluster.Settings,
+	serverCfg *execinfra.ServerConfig,
+	feedCfg jobspb.ChangefeedDetails,
 	timestampOracle timestampLowerBoundOracle,
-	makeExternalStorageFromURI cloud.ExternalStorageFromURIFactory,
 	user security.SQLUsername,
 	acc mon.BoundAccount,
 ) (Sink, error) {
-	u, err := url.Parse(sinkURI)
+	u, err := url.Parse(feedCfg.SinkURI)
 	if err != nil {
 		return nil, err
 	}
-	q := u.Query()
 
-	// Use a function here to delay creation of the sink until after we've done
-	// all the parameter verification.
-	var makeSink func() (Sink, error)
-	switch {
-	case u.Scheme == changefeedbase.SinkSchemeBuffer:
-		makeSink = func() (Sink, error) { return &bufferSink{}, nil }
-	case u.Scheme == changefeedbase.SinkSchemeNull:
-		makeSink = func() (Sink, error) { return &nullSink{}, nil }
-	case u.Scheme == changefeedbase.SinkSchemeKafka:
-		var cfg kafkaSinkConfig
-		cfg.kafkaTopicPrefix = q.Get(changefeedbase.SinkParamTopicPrefix)
-		cfg.kafkaTopicName = q.Get(changefeedbase.SinkParamTopicName)
-		q.Del(changefeedbase.SinkParamTopicPrefix)
-		q.Del(changefeedbase.SinkParamTopicName)
-		if schemaTopic := q.Get(changefeedbase.SinkParamSchemaTopic); schemaTopic != `` {
-			return nil, errors.Errorf(`%s is not yet supported`, changefeedbase.SinkParamSchemaTopic)
-		}
-		q.Del(changefeedbase.SinkParamSchemaTopic)
-		if tlsBool := q.Get(changefeedbase.SinkParamTLSEnabled); tlsBool != `` {
-			var err error
-			if cfg.tlsEnabled, err = strconv.ParseBool(tlsBool); err != nil {
-				return nil, errors.Errorf(`param %s must be a bool: %s`, changefeedbase.SinkParamTLSEnabled, err)
-			}
-		}
-		q.Del(changefeedbase.SinkParamTLSEnabled)
-		if tlsVerifyBool := q.Get(changefeedbase.SinkParamSkipTLSVerify); tlsVerifyBool != `` {
-			var err error
-			if cfg.tlsSkipVerify, err = strconv.ParseBool(tlsVerifyBool); err != nil {
-				return nil, errors.Errorf(`param %s must be a bool: %s`, changefeedbase.SinkParamSkipTLSVerify, err)
-			}
-		}
-		q.Del(changefeedbase.SinkParamSkipTLSVerify)
-		if caCertHex := q.Get(changefeedbase.SinkParamCACert); caCertHex != `` {
-			// TODO(dan): There's a straightforward and unambiguous transformation
-			// between the base 64 encoding defined in RFC 4648 and the URL variant
-			// defined in the same RFC: simply replace all `+` with `-` and `/` with
-			// `_`. Consider always doing this for the user and accepting either
-			// variant.
-			if cfg.caCert, err = base64.StdEncoding.DecodeString(caCertHex); err != nil {
-				return nil, errors.Errorf(`param %s must be base 64 encoded: %s`, changefeedbase.SinkParamCACert, err)
-			}
-		}
-		q.Del(changefeedbase.SinkParamCACert)
-		if clientCertHex := q.Get(changefeedbase.SinkParamClientCert); clientCertHex != `` {
-			if cfg.clientCert, err = base64.StdEncoding.DecodeString(clientCertHex); err != nil {
-				return nil, errors.Errorf(`param %s must be base 64 encoded: %s`, changefeedbase.SinkParamClientCert, err)
-			}
-		}
-		q.Del(changefeedbase.SinkParamClientCert)
-		if clientKeyHex := q.Get(changefeedbase.SinkParamClientKey); clientKeyHex != `` {
-			if cfg.clientKey, err = base64.StdEncoding.DecodeString(clientKeyHex); err != nil {
-				return nil, errors.Errorf(`param %s must be base 64 encoded: %s`, changefeedbase.SinkParamClientKey, err)
-			}
-		}
-		q.Del(changefeedbase.SinkParamClientKey)
-
-		saslParam := q.Get(changefeedbase.SinkParamSASLEnabled)
-		q.Del(changefeedbase.SinkParamSASLEnabled)
-		if saslParam != `` {
-			b, err := strconv.ParseBool(saslParam)
-			if err != nil {
-				return nil, errors.Wrapf(err, `param %s must be a bool:`, changefeedbase.SinkParamSASLEnabled)
-			}
-			cfg.saslEnabled = b
-		}
-		handshakeParam := q.Get(changefeedbase.SinkParamSASLHandshake)
-		q.Del(changefeedbase.SinkParamSASLHandshake)
-		if handshakeParam == `` {
-			cfg.saslHandshake = true
-		} else {
-			if !cfg.saslEnabled {
-				return nil, errors.Errorf(`%s must be enabled to configure SASL handshake behavior`, changefeedbase.SinkParamSASLEnabled)
-			}
-			b, err := strconv.ParseBool(handshakeParam)
-			if err != nil {
-				return nil, errors.Wrapf(err, `param %s must be a bool:`, changefeedbase.SinkParamSASLHandshake)
-			}
-			cfg.saslHandshake = b
-		}
-
-		cfg.saslMechanism = q.Get(changefeedbase.SinkParamSASLMechanism)
-		q.Del(changefeedbase.SinkParamSASLMechanism)
-		if cfg.saslMechanism != `` && !cfg.saslEnabled {
-			return nil, errors.Errorf(`%s must be enabled to configure SASL mechanism`, changefeedbase.SinkParamSASLEnabled)
-		}
-		if cfg.saslMechanism == `` {
-			cfg.saslMechanism = sarama.SASLTypePlaintext
-		}
-		switch cfg.saslMechanism {
-		case sarama.SASLTypeSCRAMSHA256, sarama.SASLTypeSCRAMSHA512, sarama.SASLTypePlaintext:
-		default:
-			return nil, errors.Errorf(`param %s must be one of %s, %s, or %s`,
-				changefeedbase.SinkParamSASLMechanism,
-				sarama.SASLTypeSCRAMSHA256, sarama.SASLTypeSCRAMSHA512, sarama.SASLTypePlaintext)
-		}
-
-		cfg.saslUser = q.Get(changefeedbase.SinkParamSASLUser)
-		q.Del(changefeedbase.SinkParamSASLUser)
-		cfg.saslPassword = q.Get(changefeedbase.SinkParamSASLPassword)
-		q.Del(changefeedbase.SinkParamSASLPassword)
-		if cfg.saslEnabled {
-			if cfg.saslUser == `` {
-				return nil, errors.Errorf(`%s must be provided when SASL is enabled`, changefeedbase.SinkParamSASLUser)
-			}
-			if cfg.saslPassword == `` {
-				return nil, errors.Errorf(`%s must be provided when SASL is enabled`, changefeedbase.SinkParamSASLPassword)
-			}
-		} else {
-			if cfg.saslUser != `` {
-				return nil, errors.Errorf(`%s must be enabled if a SASL user is provided`, changefeedbase.SinkParamSASLEnabled)
-			}
-			if cfg.saslPassword != `` {
-				return nil, errors.Errorf(`%s must be enabled if a SASL password is provided`, changefeedbase.SinkParamSASLEnabled)
-			}
-		}
-
-		makeSink = func() (Sink, error) {
-			return makeKafkaSink(ctx, cfg, u.Host, targets, opts, acc)
-		}
-	case isCloudStorageSink(u):
-		fileSizeParam := q.Get(changefeedbase.SinkParamFileSize)
-		q.Del(changefeedbase.SinkParamFileSize)
-		var fileSize int64 = 16 << 20 // 16MB
-		if fileSizeParam != `` {
-			if fileSize, err = humanizeutil.ParseBytes(fileSizeParam); err != nil {
-				return nil, pgerror.Wrapf(err, pgcode.Syntax, `parsing %s`, fileSizeParam)
-			}
-		}
-		u.Scheme = strings.TrimPrefix(u.Scheme, `experimental-`)
-		// Transfer "ownership" of validating all remaining query parameters to
-		// ExternalStorage.
-		u.RawQuery = q.Encode()
-		q = url.Values{}
-		makeSink = func() (Sink, error) {
+	newSink := func() (Sink, error) {
+		switch {
+		case u.Scheme == changefeedbase.SinkSchemeBuffer:
+			return &bufferSink{}, nil
+		case u.Scheme == changefeedbase.SinkSchemeNull:
+			return &nullSink{}, nil
+		case u.Scheme == changefeedbase.SinkSchemeKafka:
+			return makeKafkaSink(ctx, sinkURL{URL: u}, feedCfg.Targets, feedCfg.Opts, acc)
+		case isCloudStorageSink(u):
 			return makeCloudStorageSink(
-				ctx, u.String(), srcID, fileSize, settings,
-				opts, timestampOracle, makeExternalStorageFromURI, user, acc,
+				ctx, sinkURL{URL: u}, serverCfg.NodeID.SQLInstanceID(), serverCfg.Settings,
+				feedCfg.Opts, timestampOracle, serverCfg.ExternalStorageFromURI, user, acc,
 			)
+		case u.Scheme == changefeedbase.SinkSchemeExperimentalSQL:
+			return makeSQLSink(sinkURL{URL: u}, sqlSinkTableName, feedCfg.Targets)
+		default:
+			return nil, errors.Errorf(`unsupported sink: %s`, u.Scheme)
 		}
-	case u.Scheme == changefeedbase.SinkSchemeExperimentalSQL:
-		// Swap the changefeed prefix for the sql connection one that sqlSink
-		// expects.
-		u.Scheme = `postgres`
-		// TODO(dan): Make tableName configurable or based on the job ID or
-		// something.
-		tableName := `sqlsink`
-		makeSink = func() (Sink, error) {
-			return makeSQLSink(u.String(), tableName, targets)
-		}
-		// Remove parameters we know about for the unknown parameter check.
-		q.Del(`sslcert`)
-		q.Del(`sslkey`)
-		q.Del(`sslmode`)
-		q.Del(`sslrootcert`)
-	default:
-		return nil, errors.Errorf(`unsupported sink: %s`, u.Scheme)
 	}
 
-	for k := range q {
-		return nil, errors.Errorf(`unknown sink query parameter: %s`, k)
-	}
-
-	s, err := makeSink()
+	sink, err := newSink()
 	if err != nil {
 		return nil, err
 	}
-	return s, nil
+
+	if err := sink.Dial(); err != nil {
+		return nil, err
+	}
+
+	return sink, nil
+}
+
+// sinkURL is a helper struct which for "consuming" URL query
+// parameters from the underlying URL.
+type sinkURL struct {
+	*url.URL
+	q url.Values
+}
+
+func (u *sinkURL) consumeParam(p string) string {
+	if u.q == nil {
+		u.q = u.Query()
+	}
+	v := u.q.Get(p)
+	u.q.Del(p)
+	return v
+}
+
+func (u *sinkURL) remainingQueryParams() (res []string) {
+	for p := range u.q {
+		res = append(res, p)
+	}
+	return
+}
+
+func (u sinkURL) String() string {
+	if u.q != nil {
+		// If we changed query params, re-encode them.
+		u.URL.RawQuery = u.q.Encode()
+		u.q = nil
+	}
+	return u.URL.String()
 }
 
 // errorWrapperSink delegates to another sink and marks all returned errors as
@@ -273,6 +143,7 @@ type errorWrapperSink struct {
 	wrapped Sink
 }
 
+// EmitRow implements Sink interface.
 func (s errorWrapperSink) EmitRow(
 	ctx context.Context, topicDescr TopicDescriptor, key, value []byte, updated hlc.Timestamp,
 ) error {
@@ -282,6 +153,7 @@ func (s errorWrapperSink) EmitRow(
 	return nil
 }
 
+// EmitResolvedTimestamp implements Sink interface.
 func (s errorWrapperSink) EmitResolvedTimestamp(
 	ctx context.Context, encoder Encoder, resolved hlc.Timestamp,
 ) error {
@@ -291,6 +163,7 @@ func (s errorWrapperSink) EmitResolvedTimestamp(
 	return nil
 }
 
+// Flush implements Sink interface.
 func (s errorWrapperSink) Flush(ctx context.Context) error {
 	if err := s.wrapped.Flush(ctx); err != nil {
 		return MarkRetryableError(err)
@@ -298,6 +171,7 @@ func (s errorWrapperSink) Flush(ctx context.Context) error {
 	return nil
 }
 
+// Close implements Sink interface.
 func (s errorWrapperSink) Close() error {
 	if err := s.wrapped.Close(); err != nil {
 		return MarkRetryableError(err)
@@ -305,647 +179,9 @@ func (s errorWrapperSink) Close() error {
 	return nil
 }
 
-type kafkaLogAdapter struct {
-	ctx context.Context
-}
-
-var _ sarama.StdLogger = (*kafkaLogAdapter)(nil)
-
-func (l *kafkaLogAdapter) Print(v ...interface{}) {
-	log.InfofDepth(l.ctx, 1, "", v...)
-}
-func (l *kafkaLogAdapter) Printf(format string, v ...interface{}) {
-	log.InfofDepth(l.ctx, 1, format, v...)
-}
-func (l *kafkaLogAdapter) Println(v ...interface{}) {
-	log.InfofDepth(l.ctx, 1, "", v...)
-}
-
-func init() {
-	// We'd much prefer to make one of these per sink, so we can use the real
-	// context, but quite unfortunately, sarama only has a global logger hook.
-	ctx := context.Background()
-	ctx = logtags.AddTag(ctx, "kafka-producer", nil)
-	sarama.Logger = &kafkaLogAdapter{ctx: ctx}
-}
-
-type kafkaSinkConfig struct {
-	kafkaTopicPrefix string
-	kafkaTopicName   string
-	tlsEnabled       bool
-	tlsSkipVerify    bool
-	caCert           []byte
-	clientCert       []byte
-	clientKey        []byte
-	saslEnabled      bool
-	saslHandshake    bool
-	saslUser         string
-	saslPassword     string
-	saslMechanism    string
-}
-
-// kafkaSink emits to Kafka asynchronously. It is not concurrency-safe; all
-// calls to Emit and Flush should be from the same goroutine.
-type kafkaSink struct {
-	ctx      context.Context
-	cfg      kafkaSinkConfig
-	client   sarama.Client
-	producer sarama.AsyncProducer
-	topics   map[descpb.ID]string
-
-	lastMetadataRefresh time.Time
-
-	stopWorkerCh chan struct{}
-	worker       sync.WaitGroup
-	scratch      bufalloc.ByteAllocator
-
-	// Only synchronized between the client goroutine and the worker goroutine.
-	mu struct {
-		syncutil.Mutex
-		mem      mon.BoundAccount
-		inflight int64
-		flushErr error
-		flushCh  chan struct{}
-	}
-}
-
-func makeTopicsMap(
-	prefix string, name string, targets jobspb.ChangefeedTargets,
-) map[descpb.ID]string {
-	topics := make(map[descpb.ID]string)
-	useSingleName := (name != "")
-	if useSingleName {
-		name = prefix + SQLNameToKafkaName(name)
-	}
-	for id, t := range targets {
-		if useSingleName {
-			topics[id] = name
-		} else {
-			topics[id] = prefix + SQLNameToKafkaName(t.StatementTimeName)
-		}
-	}
-	return topics
-}
-
-type jsonDuration time.Duration
-
-func (j *jsonDuration) UnmarshalJSON(b []byte) error {
-	var s string
-	if err := json.Unmarshal(b, &s); err != nil {
-		return err
-	}
-	dur, err := time.ParseDuration(s)
-	if err != nil {
-		return err
-	}
-	*j = jsonDuration(dur)
-	return nil
-}
-
-type saramaConfig struct {
-	// These settings mirror ones in sarama config.
-	// We just tag them w/ JSON annotations.
-	// Flush describes settings specific to producer flushing.
-	// See sarama.Config.Producer.Flush
-	Flush struct {
-		Bytes       int          `json:",omitempty"`
-		Messages    int          `json:",omitempty"`
-		Frequency   jsonDuration `json:",omitempty"`
-		MaxMessages int          `json:",omitempty"`
-	}
-}
-
-// Apply configures provided kafka configuration struct based on this config.
-func (c *saramaConfig) Apply(kafka *sarama.Config) {
-	kafka.Producer.Flush.Bytes = c.Flush.Bytes
-	kafka.Producer.Flush.Messages = c.Flush.Messages
-	kafka.Producer.Flush.Frequency = time.Duration(c.Flush.Frequency)
-	kafka.Producer.Flush.MaxMessages = c.Flush.MaxMessages
-}
-
-var defaultSaramaConfig = func() *saramaConfig {
-	config := &saramaConfig{}
-
-	// When we emit messages to sarama, they're placed in a queue (as does any
-	// reasonable kafka producer client). When our sink's Flush is called, we
-	// have to wait for all buffered and inflight requests to be sent and then
-	// acknowledged. Quite unfortunately, we have no way to hint to the producer
-	// that it should immediately send out whatever is buffered. This
-	// configuration can have a dramatic impact on how quickly this happens
-	// naturally (and some configurations will block forever!).
-	//
-	// We can configure the producer to send out its batches based on number of
-	// messages and/or total buffered message size and/or time. If none of them
-	// are set, it uses some defaults, but if any of the three are set, it does
-	// no defaulting. Which means that if `Flush.Messages` is set to 10 and
-	// nothing else is set, then 9/10 times `Flush` will block forever. We can
-	// work around this by also setting `Flush.Frequency` but a cleaner way is
-	// to set `Flush.Messages` to 1. In the steady state, this sends a request
-	// with some messages, buffers any messages that come in while it is in
-	// flight, then sends those out.
-	config.Flush.Messages = 1
-
-	// This works around what seems to be a bug in sarama where it isn't
-	// computing the right value to compare against `Producer.MaxMessageBytes`
-	// and the server sends it back with a "Message was too large, server
-	// rejected it to avoid allocation" error. The other flush tunings are
-	// hints, but this one is a hard limit, so it's useful here as a workaround.
-	//
-	// This workaround should probably be something like setting
-	// `Producer.MaxMessageBytes` to 90% of it's value for some headroom, but
-	// this workaround is the one that's been running in roachtests and I'd want
-	// to test this one more before changing it.
-	config.Flush.MaxMessages = 1000
-
-	// config.Producer.Flush.Messages is set to 1 so we don't need this, but
-	// sarama prints scary things to the logs if we don't.
-	config.Flush.Frequency = jsonDuration(time.Hour)
-
-	return config
-}()
-
-func getSaramaConfig(opts map[string]string) (config *saramaConfig, err error) {
-	if configStr, haveOverride := opts[changefeedbase.OptKafkaSinkConfig]; haveOverride {
-		config = &saramaConfig{}
-		err = json.Unmarshal([]byte(configStr), config)
-	} else {
-		config = defaultSaramaConfig
-	}
-	return
-}
-
-func makeKafkaSink(
-	ctx context.Context,
-	cfg kafkaSinkConfig,
-	bootstrapServers string,
-	targets jobspb.ChangefeedTargets,
-	opts map[string]string,
-	acc mon.BoundAccount,
-) (Sink, error) {
-	sink := &kafkaSink{
-		ctx:    ctx,
-		cfg:    cfg,
-		topics: makeTopicsMap(cfg.kafkaTopicPrefix, cfg.kafkaTopicName, targets),
-	}
-	sink.mu.mem = acc
-
-	config := sarama.NewConfig()
-	config.ClientID = `CockroachDB`
-	config.Producer.Return.Successes = true
-	config.Producer.Partitioner = newChangefeedPartitioner
-
-	if cfg.caCert != nil {
-		if !cfg.tlsEnabled {
-			return nil, errors.Errorf(`%s requires %s=true`, changefeedbase.SinkParamCACert, changefeedbase.SinkParamTLSEnabled)
-		}
-		caCertPool := x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM(cfg.caCert)
-		config.Net.TLS.Config = &tls.Config{
-			RootCAs: caCertPool,
-		}
-		config.Net.TLS.Enable = true
-	} else if cfg.tlsEnabled {
-		config.Net.TLS.Enable = true
-	}
-
-	if cfg.tlsEnabled {
-		if config.Net.TLS.Config == nil {
-			config.Net.TLS.Config = &tls.Config{}
-		}
-		config.Net.TLS.Config.InsecureSkipVerify = cfg.tlsSkipVerify
-	}
-
-	if cfg.clientCert != nil {
-		if !cfg.tlsEnabled {
-			return nil, errors.Errorf(`%s requires %s=true`, changefeedbase.SinkParamClientCert, changefeedbase.SinkParamTLSEnabled)
-		}
-		if cfg.clientKey == nil {
-			return nil, errors.Errorf(`%s requires %s to be set`, changefeedbase.SinkParamClientCert, changefeedbase.SinkParamClientKey)
-		}
-		cert, err := tls.X509KeyPair(cfg.clientCert, cfg.clientKey)
-		if err != nil {
-			return nil, errors.Errorf(`invalid client certificate data provided: %s`, err)
-		}
-		config.Net.TLS.Config.Certificates = []tls.Certificate{cert}
-	} else if cfg.clientKey != nil {
-		return nil, errors.Errorf(`%s requires %s to be set`, changefeedbase.SinkParamClientKey, changefeedbase.SinkParamClientCert)
-	}
-
-	if cfg.saslEnabled {
-		config.Net.SASL.Enable = true
-		config.Net.SASL.Handshake = cfg.saslHandshake
-		config.Net.SASL.User = cfg.saslUser
-		config.Net.SASL.Password = cfg.saslPassword
-		config.Net.SASL.Mechanism = sarama.SASLMechanism(cfg.saslMechanism)
-		switch config.Net.SASL.Mechanism {
-		case sarama.SASLTypeSCRAMSHA512:
-			config.Net.SASL.SCRAMClientGeneratorFunc = sha512ClientGenerator
-		case sarama.SASLTypeSCRAMSHA256:
-			config.Net.SASL.SCRAMClientGeneratorFunc = sha256ClientGenerator
-		}
-	}
-
-	saramaCfg, err := getSaramaConfig(opts)
-	if err != nil {
-		return nil, errors.Wrapf(err,
-			"failed to parse sarama config; check %s option", changefeedbase.OptKafkaSinkConfig)
-	}
-	saramaCfg.Apply(config)
-
-	sink.client, err = sarama.NewClient(strings.Split(bootstrapServers, `,`), config)
-	if err != nil {
-		err = pgerror.Wrapf(err, pgcode.CannotConnectNow,
-			`connecting to kafka: %s`, bootstrapServers)
-		return nil, err
-	}
-	sink.producer, err = sarama.NewAsyncProducerFromClient(sink.client)
-	if err != nil {
-		err = pgerror.Wrapf(err, pgcode.CannotConnectNow,
-			`connecting to kafka: %s`, bootstrapServers)
-		return nil, err
-	}
-
-	sink.start()
-	return sink, nil
-}
-
-func (s *kafkaSink) start() {
-	s.stopWorkerCh = make(chan struct{})
-	s.worker.Add(1)
-	go s.workerLoop()
-}
-
-// Close implements the Sink interface.
-func (s *kafkaSink) Close() error {
-	defer func() {
-		s.mu.Lock()
-		s.mu.mem.Close(s.ctx)
-		s.mu.Unlock()
-	}()
-
-	close(s.stopWorkerCh)
-	s.worker.Wait()
-	// If we're shutting down, we don't care what happens to the outstanding
-	// messages, so ignore this error.
-	_ = s.producer.Close()
-	// s.client is only nil in tests.
-	if s.client != nil {
-		return s.client.Close()
-	}
-	return nil
-}
-
-// EmitRow implements the Sink interface.
-func (s *kafkaSink) EmitRow(
-	ctx context.Context, topicDescr TopicDescriptor, key, value []byte, updated hlc.Timestamp,
-) error {
-	topic, isKnownTopic := s.topics[topicDescr.GetID()]
-	if !isKnownTopic {
-		return errors.Errorf(`cannot emit to undeclared topic: %s`, topicDescr.GetName())
-	}
-
-	msg := &sarama.ProducerMessage{
-		Topic: topic,
-		Key:   sarama.ByteEncoder(key),
-		Value: sarama.ByteEncoder(value),
-	}
-	return s.emitMessage(ctx, msg)
-}
-
-// EmitResolvedTimestamp implements the Sink interface.
-func (s *kafkaSink) EmitResolvedTimestamp(
-	ctx context.Context, encoder Encoder, resolved hlc.Timestamp,
-) error {
-	// Periodically ping sarama to refresh its metadata. This means talking to
-	// zookeeper, so it shouldn't be done too often, but beyond that this
-	// constant was picked pretty arbitrarily.
-	//
-	// TODO(dan): Add a test for this. We can't right now (2018-11-13) because
-	// we'd need to bump sarama, but that's a bad idea while we're still
-	// actively working on stability. At the same time, revisit this tuning.
-	const metadataRefreshMinDuration = time.Minute
-	if timeutil.Since(s.lastMetadataRefresh) > metadataRefreshMinDuration {
-		topics := make([]string, 0, len(s.topics))
-		for _, topic := range s.topics {
-			topics = append(topics, topic)
-		}
-		if err := s.client.RefreshMetadata(topics...); err != nil {
-			return err
-		}
-		s.lastMetadataRefresh = timeutil.Now()
-	}
-
-	for _, topic := range s.topics {
-		payload, err := encoder.EncodeResolvedTimestamp(ctx, topic, resolved)
-		if err != nil {
-			return err
-		}
-		s.scratch, payload = s.scratch.Copy(payload, 0 /* extraCap */)
-
-		// sarama caches this, which is why we have to periodically refresh the
-		// metadata above. Staleness here does not impact correctness. Some new
-		// partitions will miss this resolved timestamp, but they'll eventually
-		// be picked up and get later ones.
-		partitions, err := s.client.Partitions(topic)
-		if err != nil {
-			return err
-		}
-		for _, partition := range partitions {
-			msg := &sarama.ProducerMessage{
-				Topic:     topic,
-				Partition: partition,
-				Key:       nil,
-				Value:     sarama.ByteEncoder(payload),
-			}
-			if err := s.emitMessage(ctx, msg); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Flush implements the Sink interface.
-func (s *kafkaSink) Flush(ctx context.Context) error {
-	flushCh := make(chan struct{}, 1)
-
-	s.mu.Lock()
-	inflight := s.mu.inflight
-	flushErr := s.mu.flushErr
-	s.mu.flushErr = nil
-	immediateFlush := inflight == 0 || flushErr != nil
-	if !immediateFlush {
-		s.mu.flushCh = flushCh
-	}
-	s.mu.Unlock()
-
-	if immediateFlush {
-		return flushErr
-	}
-
-	if log.V(1) {
-		log.Infof(ctx, "flush waiting for %d inflight messages", inflight)
-	}
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-flushCh:
-		s.mu.Lock()
-		flushErr := s.mu.flushErr
-		s.mu.flushErr = nil
-		s.mu.Unlock()
-		return flushErr
-	}
-}
-
-func kafkaMessageBytes(m *sarama.ProducerMessage) (s int64) {
-	if m.Key != nil {
-		s += int64(m.Key.Length())
-	}
-	if m.Value != nil {
-		s += int64(m.Value.Length())
-	}
-	return
-}
-
-func (s *kafkaSink) startInflightMessage(ctx context.Context, msg *sarama.ProducerMessage) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if err := s.mu.mem.Grow(ctx, kafkaMessageBytes(msg)); err != nil {
-		return err
-	}
-
-	s.mu.inflight++
-
-	if log.V(2) {
-		log.Infof(ctx, "emitting %d inflight records to kafka", s.mu.inflight)
-	}
-	return nil
-}
-
-func (s *kafkaSink) emitMessage(ctx context.Context, msg *sarama.ProducerMessage) error {
-	if err := s.startInflightMessage(ctx, msg); err != nil {
-		return err
-	}
-
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case s.producer.Input() <- msg:
-	}
-
-	return nil
-}
-
-func (s *kafkaSink) workerLoop() {
-	defer s.worker.Done()
-
-	for {
-		var ackMsg *sarama.ProducerMessage
-		var ackError error
-
-		select {
-		case <-s.stopWorkerCh:
-			return
-		case m := <-s.producer.Successes():
-			ackMsg = m
-		case err := <-s.producer.Errors():
-			ackMsg, ackError = err.Msg, err.Err
-		}
-
-		s.mu.Lock()
-		s.mu.inflight--
-		s.mu.mem.Shrink(s.ctx, kafkaMessageBytes(ackMsg))
-		if s.mu.flushErr == nil && ackError != nil {
-			s.mu.flushErr = ackError
-		}
-
-		if s.mu.inflight == 0 && s.mu.flushCh != nil {
-			s.mu.flushCh <- struct{}{}
-			s.mu.flushCh = nil
-		}
-		s.mu.Unlock()
-	}
-}
-
-type changefeedPartitioner struct {
-	hash sarama.Partitioner
-}
-
-var _ sarama.Partitioner = &changefeedPartitioner{}
-var _ sarama.PartitionerConstructor = newChangefeedPartitioner
-
-func newChangefeedPartitioner(topic string) sarama.Partitioner {
-	return &changefeedPartitioner{
-		hash: sarama.NewHashPartitioner(topic),
-	}
-}
-
-func (p *changefeedPartitioner) RequiresConsistency() bool { return true }
-func (p *changefeedPartitioner) Partition(
-	message *sarama.ProducerMessage, numPartitions int32,
-) (int32, error) {
-	if message.Key == nil {
-		return message.Partition, nil
-	}
-	return p.hash.Partition(message, numPartitions)
-}
-
-const (
-	sqlSinkCreateTableStmt = `CREATE TABLE IF NOT EXISTS "%s" (
-		topic STRING,
-		partition INT,
-		message_id INT,
-		key BYTES, value BYTES,
-		resolved BYTES,
-		PRIMARY KEY (topic, partition, message_id)
-	)`
-	sqlSinkEmitStmt = `INSERT INTO "%s" (topic, partition, message_id, key, value, resolved)`
-	sqlSinkEmitCols = 6
-	// Some amount of batching to mirror a bit how kafkaSink works.
-	sqlSinkRowBatchSize = 3
-	// While sqlSink is only used for testing, hardcode the number of
-	// partitions to something small but greater than 1.
-	sqlSinkNumPartitions = 3
-)
-
-// sqlSink mirrors the semantics offered by kafkaSink as closely as possible,
-// but writes to a SQL table (presumably in CockroachDB). Currently only for
-// testing.
-//
-// Each emitted row or resolved timestamp is stored as a row in the table. Each
-// table gets 3 partitions. Similar to kafkaSink, the order between two emits is
-// only preserved if they are emitted to by the same node and to the same
-// partition.
-type sqlSink struct {
-	db *gosql.DB
-
-	tableName string
-	topics    map[string]struct{}
-	hasher    hash.Hash32
-
-	rowBuf  []interface{}
-	scratch bufalloc.ByteAllocator
-
-	targetNames map[descpb.ID]string
-}
-
-func makeSQLSink(uri, tableName string, targets jobspb.ChangefeedTargets) (*sqlSink, error) {
-	if u, err := url.Parse(uri); err != nil {
-		return nil, err
-	} else if u.Path == `` {
-		return nil, errors.Errorf(`must specify database`)
-	}
-	db, err := gosql.Open(`postgres`, uri)
-	if err != nil {
-		return nil, err
-	}
-	if _, err := db.Exec(fmt.Sprintf(sqlSinkCreateTableStmt, tableName)); err != nil {
-		db.Close()
-		return nil, err
-	}
-
-	s := &sqlSink{
-		db:          db,
-		tableName:   tableName,
-		topics:      make(map[string]struct{}),
-		hasher:      fnv.New32a(),
-		targetNames: make(map[descpb.ID]string),
-	}
-	for id, t := range targets {
-		s.topics[t.StatementTimeName] = struct{}{}
-		s.targetNames[id] = t.StatementTimeName
-	}
-	return s, nil
-}
-
-// EmitRow implements the Sink interface.
-func (s *sqlSink) EmitRow(
-	ctx context.Context, topicDescr TopicDescriptor, key, value []byte, updated hlc.Timestamp,
-) error {
-	topic := s.targetNames[topicDescr.GetID()]
-	if _, ok := s.topics[topic]; !ok {
-		return errors.Errorf(`cannot emit to undeclared topic: %s`, topic)
-	}
-
-	// Hashing logic copied from sarama.HashPartitioner.
-	s.hasher.Reset()
-	if _, err := s.hasher.Write(key); err != nil {
-		return err
-	}
-	partition := int32(s.hasher.Sum32()) % sqlSinkNumPartitions
-	if partition < 0 {
-		partition = -partition
-	}
-
-	var noResolved []byte
-	return s.emit(ctx, topic, partition, key, value, noResolved)
-}
-
-// EmitResolvedTimestamp implements the Sink interface.
-func (s *sqlSink) EmitResolvedTimestamp(
-	ctx context.Context, encoder Encoder, resolved hlc.Timestamp,
-) error {
-	var noKey, noValue []byte
-	for topic := range s.topics {
-		payload, err := encoder.EncodeResolvedTimestamp(ctx, topic, resolved)
-		if err != nil {
-			return err
-		}
-		s.scratch, payload = s.scratch.Copy(payload, 0 /* extraCap */)
-		for partition := int32(0); partition < sqlSinkNumPartitions; partition++ {
-			if err := s.emit(ctx, topic, partition, noKey, noValue, payload); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func (s *sqlSink) emit(
-	ctx context.Context, topic string, partition int32, key, value, resolved []byte,
-) error {
-	// Generate the message id on the client to match the guaranttees of kafka
-	// (two messages are only guaranteed to keep their order if emitted from the
-	// same producer to the same partition).
-	messageID := builtins.GenerateUniqueInt(base.SQLInstanceID(partition))
-	s.rowBuf = append(s.rowBuf, topic, partition, messageID, key, value, resolved)
-	if len(s.rowBuf)/sqlSinkEmitCols >= sqlSinkRowBatchSize {
-		return s.Flush(ctx)
-	}
-	return nil
-}
-
-// Flush implements the Sink interface.
-func (s *sqlSink) Flush(ctx context.Context) error {
-	if len(s.rowBuf) == 0 {
-		return nil
-	}
-
-	var stmt strings.Builder
-	fmt.Fprintf(&stmt, sqlSinkEmitStmt, s.tableName)
-	for i := 0; i < len(s.rowBuf); i++ {
-		if i == 0 {
-			stmt.WriteString(` VALUES (`)
-		} else if i%sqlSinkEmitCols == 0 {
-			stmt.WriteString(`),(`)
-		} else {
-			stmt.WriteString(`,`)
-		}
-		fmt.Fprintf(&stmt, `$%d`, i+1)
-	}
-	stmt.WriteString(`)`)
-	_, err := s.db.Exec(stmt.String(), s.rowBuf...)
-	if err != nil {
-		return err
-	}
-	s.rowBuf = s.rowBuf[:0]
-	return nil
-}
-
-// Close implements the Sink interface.
-func (s *sqlSink) Close() error {
-	return s.db.Close()
+// Dial implements Sink interface.
+func (s errorWrapperSink) Dial() error {
+	return s.wrapped.Dial()
 }
 
 // encDatumRowBuffer is a FIFO of `EncDatumRow`s.
@@ -1022,9 +258,16 @@ func (s *bufferSink) Close() error {
 	return nil
 }
 
-type nullSink struct {
+// Dial implements the Sink interface.
+func (s *bufferSink) Dial() error {
+	return nil
 }
 
+type nullSink struct{}
+
+var _ Sink = (*nullSink)(nil)
+
+// EmitRow implements SInk interface.
 func (n *nullSink) EmitRow(
 	ctx context.Context, topic TopicDescriptor, key, value []byte, updated hlc.Timestamp,
 ) error {
@@ -1032,6 +275,7 @@ func (n *nullSink) EmitRow(
 	return nil
 }
 
+// EmitResolvedTimestamp implements SInk interface.
 func (n *nullSink) EmitResolvedTimestamp(
 	ctx context.Context, encoder Encoder, resolved hlc.Timestamp,
 ) error {
@@ -1042,6 +286,7 @@ func (n *nullSink) EmitResolvedTimestamp(
 	return nil
 }
 
+// Flush implements SInk interface.
 func (n *nullSink) Flush(ctx context.Context) error {
 	if log.V(2) {
 		log.Info(ctx, "flushing")
@@ -1050,8 +295,12 @@ func (n *nullSink) Flush(ctx context.Context) error {
 	return nil
 }
 
+// Close implements SInk interface.
 func (n *nullSink) Close() error {
 	return nil
 }
 
-var _ Sink = (*nullSink)(nil)
+// Dial implements SInk interface.
+func (n *nullSink) Dial() error {
+	return nil
+}
